Model View Controller in Express
