+++ controllers/messages.controller.js
@@ -1,0 +1,12 @@
+function getMessages(req,res) {
+    res.send('<ul><li>Hellooooworld</li></ul>');
+}
+
+function postMessage(req,res){
+    console.log('Updating messages...');
+}
+
+module.exports = {
+    getMessages,
+    postMessage
+}
