+++ models/friends.model.js
@@ -1,0 +1,16 @@
+const friends = [
+    {  
+    id: 0,
+    name: "Linus Torvalds"
+    },
+    {
+    id: 1,
+    name: "Edward Diaz"
+    },
+    {
+    id: 2,
+    name: "Issac Newton"
+    }
+];
+
+module.exports = friends;
