+++ server.js
@@ -1,23 +1,12 @@
 const express = require('express');
+const messagesControllers = require ('./controllers/messages.controller');
+const friendsControllers = require ('./controllers/friends.controller');
 
 const app = express();
 
 const PORT = 3000;
 
-const friends = [
-    {  
-    id: 0,
-    name: "Linus Torvalds"
-    },
-    {
-    id: 1,
-    name: "Edward Diaz"
-    },
-    {
-    id: 2,
-    name: "Issac Newton"
-    }
-];
+
 
 app.use((req,res,next)=>{
     const start = Date.now();
@@ -28,44 +17,12 @@ app.use((req,res,next)=>{
 
 app.use(express.json());
 
-app.post('/friends',(req, res) =>{
-   if (!req.body.name) {
-    return res.status(400).json({
-        error: 'Missing friend name'
-    })
-   }
-    const newFriend = {
-        name: req.body.name,
-        id: friends.length
-    };
-    friends.push(newFriend);
-    res.json(newFriend);
-});
+app.post('/friends',friendsControllers.postFriends);
+app.get('/friends',friendsControllers.getFriends );
+app.get('/friends/:friendId',friendsControllers.getFriendsbyId );
 
-app.get('/friends', (req,res) =>{
-    res.json(friends);
-});
-
-app.get('/friends/:friendId', (req,res) =>{
-    const friendId = Number(req.params.friendId);
-    const friend = friends[friendId];
-    if (friend) {
-        res.status(200).json(friend);
-    }else {
-        res.status(404).json({
-            error: "Friend not found"
-        })
-    }
-});
-
-app.get('/messages', (req,res) =>{
-    res.send('<ul><li>Hellooooworld</li></ul>');
-});
-
-
-app.post('/messages', (req,res) =>{
-    console.log('Updating messages...');
-});
+app.get('/messages',messagesControllers.getMessages);
+app.post('/messages',messagesControllers.postMessage);
 
 app.listen(PORT,() => {
     console.log(`Listening on port ${PORT}...`);
